requirements.txt
Flask==2.1.1
Flask-MySQLdb==0.2.0
SQLAlchemy==1.4.32
mysql-connector-python==8.0.28
